+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 If no more bugs are found, this version will become the release version.
 
-* Fixed the check for enabled datapacks, which previously resulted in all datapacks being enabled during the first world reload.
-* Added more Minecraft versions.
+* Fixed a bug with the /datapack disable command, which caused nothing to happen
+* Datapacks are now copied to the world directory without using the /reload command immediately after /datapack enable
+* Fixed a bug where the server would use optional datapacks when created